requirements.txt
numpy==1.23.0
pandas==1.4.3
openpyxl==3.0.10
mysql-connector-python==8.0.31
